♻️ refactor(embeddings): Store embeddings as base64 encoded text
This refactors the internal representation of vector embeddings within the `Point` model from a list of floats to a base64 encoded string. This change improves storage efficiency and optimizes data handling for DuckDB. A `to_python()` method is introduced on the `Point` object for converting the encoded string back to a float list. The default batch size for creating points has also been increased.